Revert "Merge remote-tracking branch 'origin/releasecandidate/v1.0.0' into copilot/fix-post-response-capacity"
This reverts commit 69aed7ef91adf8b66621940324efb8c948c3c384, reversing
changes made to f336a8fe3369ed3c3d7c2d371f9977a9caa568ea.

+++ Poliedro.Eds.Application/Compartiment/Commands/UpdateServer/UpdateCompartimentCommand.cs
@@ -7,8 +7,8 @@ namespace Poliedro.Eds.Application.Compartiment.Commands.UpdateCompartiment;
 public record UpdateCompartimentCommand(
 int IdCompartment,
 int Number,
-decimal Nominal,
-decimal Operative,
+double Nominal,
+double Operative,
 int IdProduct,
-decimal Height,
+double Height,
 int IdTank) : IRequest<Result<VoidResult, Error>>;
